requirements.txt
selenium==4.15.1
python-dotenv==1.0.1
colorama==0.4.6
PyYAML~=5.3.1
pynput==1.7.6